Created simple classes for world representation.
TerrainTile is data representing single terrain tile in the world, for now it only contains Name (type) of the tile.

World hold a dictionary of these tiles (dictionary to allow negative indexes) and right now is initialized randomly with 4 possible type types.

WorldNode is scene to visually represent the world, for now it only contains one tile layer to display the terrain from the provided wrold data.

Added simple tile texture atlas with 2 alternatives for each of the 4 terrains.

Set texture filter to linear to prevent blending of colors.

Added the utility classes for loading and instantiating scenes without the need to explicitly define their paths. To make sure everything works as intended, scenes need to be in Scenes folder while scripts need to be in Scripts folder and the subfolder structure needs to be identical in both. Path of the .cs and its .tscn need to match except for the Scripts/Scenes part.