Minor fixes.
A little rework of SubBiome function. The envitomant variables are now accessed in the method and the position is the initial tile position, not the island offset position. Also moved the sub biome dictionary to conf (has to be Godot.Collections.Dictionary not the native C# dict to be exported) so all config is in one place.

Fixed some comments.